+++ src/product_listing/index.html
@@ -5,8 +5,8 @@
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Sleep Outside | Home</title>
-  <link rel="stylesheet" href="/css/style.css" />
-  <script src="js/main.js" type="module"></script>
+  <link rel="stylesheet" href="../css/style.css" />
+  <script src="../js/product-listing.js" type="module"></script>
 </head>
 
 <body>
@@ -39,32 +39,7 @@
     </div>
   </header>
   <main class="divider">
-    <div class="hero">
-      <img src="/images/banner-sm.jpg" srcset="images/banner-sm.jpg 500w, images/banner.jpg"
-        alt="image of a high mountain lake" />
-      <div class="logo logo--square">
-        <img src="/images/noun_Tent_2517.svg" alt="tent image for logo" />
-        <div>Sleep<span class="highlight">Outside</span></div>
-      </div>
-    </div>
-    <div class="product-listing">
-      <a href="./product_listing/index.html?category=tents">
-        <img class="product-icon" src="./public/images/product-icons/category-tents.svg" alt="category-tents">
-      </a>
-    </div>
-    <section class="mission">
-      <p>We spend too much time these days indoors...sitting usually.</p>
-      <p>No wonder we are a mess!</p>
-      <p>
-        What we really need is to get outside...in fact SLEEP outside. Our
-        mission is to provide you with everything you need to do this
-        comfortably at an affordable price.
-      </p>
-    </section>
-    <section class="products">
-      <h2>Top Products</h2>
-      <ul class="product-list"></ul>
-    </section>
+    
   </main>
   <footer>&copy;NOT a real business</footer>
 </body>
