+++ vite.config.js
@@ -14,6 +14,9 @@ export default defineConfig({
         product: resolve(__dirname, "src/product_pages/index.html"
         ),
 
+        product_listing: resolve(__dirname, "src/product_listing/index.html"
+        ),
+
       },
     },
   },
